==> 3_back_part.ipynb <==
import pymysql
from sqlalchemy import create_engine
import pandas as pd
from pandas.io import sql
import datetime
import pandas as pd
import pandasql as ps
from concurrent.futures import ThreadPoolExecutor, as_completed

# reading 
now = datetime.datetime.now() 
print('시작:', now)
agg3 =  pd.read_csv('./mid/agg3.txt', sep='\t')

now = datetime.datetime.now()
print('종료:', now)

# reading 
now = datetime.datetime.now() 
print('시작:', now)

merged_references_991_to_2028 = pd.read_csv('./reference/merged_references_991_to_2028.txt',  sep='\t')
now = datetime.datetime.now()
print('종료:', now)

agg3

merged_references_991_to_2028

# 각 컬럼별 null 값의 개수 계산
null_counts000 = merged_references_991_to_2028.isnull().sum()

# 결과 출력
print(null_counts000)

조인하기

now = datetime.datetime.now() 
print('시작:', now)

third_result4 = pd.merge(merged_references_991_to_2028,agg3, on='uid', how='left')

now = datetime.datetime.now()
print('종료:', now)

# 각 컬럼별 null 값의 개수 계산
null_counts4 = third_result4.isnull().sum()

# 결과 출력
print(null_counts4)

now = datetime.datetime.now() 
print('시작:', now)

# 컬럼명 변경
third_result4 = third_result4.rename(columns={'ref_uid': 'refs_uid', 'ref_count': 'refs_n'})


# 컬럼 순서를 변경하고 싶은 순서대로 지정
new_order = ['uid' , 'cluster_id','pubyear2', 'refs_uid', 'refs_n']

# 새 순서로 DataFrame을 재구성
third_result4 = third_result4[new_order]

# 정수화
third_result4['cluster_id'] = third_result4['cluster_id'].astype('Int64')
third_result4['pubyear2'] = third_result4['pubyear2'].astype('Int64')

now = datetime.datetime.now()
print('종료:', now)

third_result4

ut_to_uid_is_core_set.pkl => P에서 ut를 통해 ut_to_uid(=uid)를 만듦 -> core가 1인 ut_to_uid(uid)만 추출한 set임.
merged_references의 uid가 이 pkl에 있다면(P에 존재하고 CORE가 1임) within_core 가 1, 아님 0으로 값 조정.

import pickle

# pickle 파일로부터 set 데이터 불러오기
with open('./mid/ut_to_uid_is_core_set.pkl', 'rb') as file:
    loaded_is_core_set = pickle.load(file)

#print(loaded_ut_ut_to_uid_is_core_set)
print(f"ut_to_uid 컬럼 데이터의 set 크기: {len(loaded_is_core_set)}")

refs_uid에 결측값이 존재함. 0과 1을 정수형으로 저장할 수 없음. 따라서 within_core의 데이터타입을 문자열로 지정함. 

이 코드를 통해 각 행에 대해 refs_uid 값이 결측치인 경우 'NaN'을, loaded_is_core_set에 포함되어 있는 경우 '1'을, 그렇지 않은 경우 '0'을 within_core 컬럼에 할당합니다

import numpy as np


now = datetime.datetime.now() 
print('시작:', now)

# third_result2 DataFrame에 within_core 컬럼 추가 및 값 할당, 문자열로 처리
third_result4['within_core'] = third_result4['refs_uid'].apply(lambda x: 'NaN' if pd.isnull(x) else ('1' if x in loaded_is_core_set else '0'))

now = datetime.datetime.now()
print('종료:', now)

 MemoryError: Unable to allocate 7.73 GiB for an array with shape (1037622925,) and data type uint64 오류는 파이썬이 요청된 크기의 배열을 메모리에 할당할 수 없을 때 발생합니다. 이 오류는 대규모 데이터를 처리하려고 할 때 특히 빈번하게 발생하며, 시스템의 사용 가능한 RAM이 요구되는 메모리 양보다 적을 때 나타납니다. 

따라서 df를 반으로 쪼개서 작업을 시도하려했으나, 위 코드 다시 돌리니 돌아가서 그대로 사용.

within_core의 데이터 값 조사, 카운트 

now = datetime.datetime.now() 
print('시작:', now)


# within_core 컬럼의 데이터 값 조사 및 카운트
within_core_counts2 = third_result4['within_core'].value_counts()

# 결과 출력
print(within_core_counts2)

now = datetime.datetime.now()
print('종료:', now)

third_result4

저장하기 -> final폴더에 .

#오래걸릴듯...
now = datetime.datetime.now() 
print('시작:', now)

third_result4.to_csv('./mid/third/third_result4.txt', sep='\t', index=False)

now = datetime.datetime.now()
print('종료:', now)

*< 작업 시 느낀점 >
1. 비동기 쓰레드를 사용했지만, DB 요청하는 파이썬 관련 작업만 빨라질 뿐 DB에서 데이터 불러오는 쿼리 작업은 병렬처리가 안되는거같음. -> 이 부분은 주피터 노트북 워크시트를 offset 시작점을 다르게하여 여러개 돌리는거로 추출 속도 보완함.  
2. limit 최적화 샘플링 -> 백만이나 십만이나 별차이 없어서 백만으로 함.그냥 DB에서 데이터 뽑고 로드하는게 오래걸리는거같음.  
*더 자세한건 보고서에 담음.
