==> 4_worksheet.ipynb <==
import pymysql
from sqlalchemy import create_engine
import pandas as pd
from pandas.io import sql
import datetime
import pandas as pd
import pandasql as ps

# reading 
now = datetime.datetime.now() 
print('시작:', now)
df =  pd.read_csv('./dict/category_subject.txt', sep='\t')
now = datetime.datetime.now()
print('종료:', now)

# 불필요 컬럼 드롭
df.drop(['cate_n'], axis=1, inplace=True)

# reading 
now = datetime.datetime.now() 
print('시작:', now)
third =  pd.read_csv('./final/third.txt', sep=',')
now = datetime.datetime.now()
print('종료:', now)

한번에 처리하려니 메모리 에러로 아래와같이 3개로 나누어 전처리


# 청크 사이즈 정의
chunk_size = 50000000

# 처리 시작 시간
total_start_time = datetime.datetime.now()

# 청크 파일 저장을 위한 리스트
chunk_files = []

# 이미 읽어온 데이터프레임을 청크로 나누어 처리
total_length = len(df)
i = 0
for start in range(0, total_length, chunk_size):
    # 청크 처리 시작 시간
    chunk_start_time = datetime.datetime.now()
    
    end = min(start + chunk_size, total_length)
    # 남은 데이터 양 출력
    print(f"처리 중: {start+1} ~ {end} / {total_length}")
    
    # 청크 분할
    df_chunk = df.iloc[start:end]
    
    # 청크와 다른 데이터셋 조인
    #result_chunk = pd.merge(df_chunk, third_df, on='uid', how='inner')
    result_chunk = pd.merge(df_chunk, third, left_on='uid', right_on='refs_uid', how='inner')
    
    # 결과 저장 파일명 정의
    chunk_file = f'./mid/four/result_chunk_{i}.txt'
    chunk_files.append(chunk_file)

    # 청크 결과를 CSV 파일로 저장
    if i == 0:
        result_chunk.to_csv(chunk_file, mode='w', index=False, header=True)  # 첫 번째 파일에만 헤더 포함
    else:
        result_chunk.to_csv(chunk_file, mode='w', index=False, header=False)  # 나머지 파일에는 헤더 미포함
    
    # 청크 처리 완료 시간 및 소요 시간 출력
    chunk_end_time = datetime.datetime.now()
    chunk_duration = chunk_end_time - chunk_start_time
    print(f"청크 {i} 처리 완료: 소요 시간 = {chunk_duration}, 남은 데이터 양 = {total_length - end}")
    
    i += 1

# 처리 종료 시간 및 전체 소요 시간 출력
total_end_time = datetime.datetime.now()
total_duration = total_end_time - total_start_time
print(f"전체 처리 완료: 소요 시간 = {total_duration}")

# 병합된 파일 생성
#output_file = './mid/four/result_combined.txt'
#with open(output_file, 'w', newline='') as f_out:
#    for file in chunk_files:
#        with open(file, 'r', newline='') as f_in:
#            f_out.write(f_in.read())

처리 중: 1 ~ 50000000 / 138594301
청크 0 처리 완료: 소요 시간 = 0:29:30.027392, 남은 데이터 양 = 88594301
처리 중: 50000001 ~ 100000000 / 138594301
청크 1 처리 완료: 소요 시간 = 0:55:50.076194, 남은 데이터 양 = 38594301
처리 중: 100000001 ~ 138594301 / 138594301
청크 2 처리 완료: 소요 시간 = 0:21:12.056753, 남은 데이터 양 = 0
전체 처리 완료: 소요 시간 = 1:46:32.161339

print(chunk_files )

['./mid/four/result_chunk_0.txt', './mid/four/result_chunk_1.txt', './mid/four/result_chunk_2.txt']

# 병합된 파일 생성
#output_file = './mid/four/result_combined.txt'
#with open(output_file, 'w', newline='') as f_out:
#    for file in chunk_files:
#        with open(file, 'r', newline='') as f_in:
#            f_out.write(f_in.read())

# reading 
#now = datetime.datetime.now() 
#print('시작:', now)
#df =  pd.read_csv('./mid/four/result_combined.txt', sep=',')
#now = datetime.datetime.now()
#print('종료:', now)

컴바인한 파일이 너무 커서 작업 불가 -> 모든 작업 후 합치는 방향으로 변경.

## 메모리 때문에 df1, df2, df3 각각 전처리 후 저장, 마지막에 최종파일로 병합.

# reading 
now = datetime.datetime.now() 
print('시작:', now)
df1 =  pd.read_csv('./mid/four/result_chunk_0.txt', sep=',', header=0)
now = datetime.datetime.now()
print('종료:', now)

header = df1.columns.tolist()

df1

# reading 
now = datetime.datetime.now() 
print('시작:', now)
df2 =  pd.read_csv('./mid/four/result_chunk_1.txt', sep=',', header=None, names=header)
now = datetime.datetime.now()
print('종료:', now)

# reading 
now = datetime.datetime.now() 
print('시작:', now)
df3 =  pd.read_csv('./mid/four/result_chunk_2.txt', sep=',', header=None, names=header)
now = datetime.datetime.now()
print('종료:', now)

df2

df3


최종 파일로 전처리 불가. 메모리 에러. 
아래와 같이 두개의 파일로 분할하여 작업 후 파일 병합.

df들을 컬럼 전처리(조인 컬럼 하나 삭제. uid_x(ref_uid)를 드롭, uid_y를 uid로 만들기.컬럼 순서변경. subject_traditinal을 cate로 변경. pubyear2를 pubyear로.

# 불필요 컬럼 드롭
df1.drop(['uid_x'], axis=1, inplace=True)
#df2.drop(['uid_x'], axis=1, inplace=True)
#df3.drop(['uid_x'], axis=1, inplace=True)
df1.rename(columns={'uid_y': 'uid', 'subject_traditional' : 'cate', 'pubyear2' : 'pubyear'}, inplace=True)
#df2.rename(columns={'uid_y': 'uid', 'subject_traditional' : 'cate', 'pubyear2' : 'pubyear'}, inplace=True)
#df3.rename(columns={'uid_y': 'uid', 'subject_traditional' : 'cate', 'pubyear2' : 'pubyear'}, inplace=True)



# 불필요 컬럼 드롭
#df1.drop(['uid_x'], axis=1, inplace=True)
#df2.drop(['uid_x'], axis=1, inplace=True)
df3.drop(['uid_x'], axis=1, inplace=True)
#df1.rename(columns={'uid_y': 'uid', 'subject_traditional' : 'cate', 'pubyear2' : 'pubyear'}, inplace=True)
#df2.rename(columns={'uid_y': 'uid', 'subject_traditional' : 'cate', 'pubyear2' : 'pubyear'}, inplace=True)
df3.rename(columns={'uid_y': 'uid', 'subject_traditional' : 'cate', 'pubyear2' : 'pubyear'}, inplace=True)



# 불필요 컬럼 드롭
#df1.drop(['uid_x'], axis=1, inplace=True)
df2.drop(['uid_x'], axis=1, inplace=True)
#df3.drop(['uid_x'], axis=1, inplace=True)
#df1.rename(columns={'uid_y': 'uid', 'subject_traditional' : 'cate', 'pubyear2' : 'pubyear'}, inplace=True)
df2.rename(columns={'uid_y': 'uid', 'subject_traditional' : 'cate', 'pubyear2' : 'pubyear'}, inplace=True)
#df3.rename(columns={'uid_y': 'uid', 'subject_traditional' : 'cate', 'pubyear2' : 'pubyear'}, inplace=True)



# 각 컬럼별 null 값의 개수 계산
null_counts22 =df2.isnull().sum()

print(null_counts22)

# 각 컬럼별 null 값의 개수 계산
null_counts33 = df3.isnull().sum()

print(null_counts33)

df1.rename(columns={'cate' : 'subject_traditional'},inplace=True)
#df2.rename(columns={'cate' : 'subject_traditional'},inplace=True)
#df3.rename(columns={'cate' : 'subject_traditional'},inplace=True)


#df1.rename(columns={'cate' : 'subject_traditional'},inplace=True)
#df2.rename(columns={'cate' : 'subject_traditional'},inplace=True)
df3.rename(columns={'cate' : 'subject_traditional'},inplace=True)


#df1.rename(columns={'cate' : 'subject_traditional'},inplace=True)
df2.rename(columns={'cate' : 'subject_traditional'},inplace=True)
#df3.rename(columns={'cate' : 'subject_traditional'},inplace=True)


cate_n생성

now = datetime.datetime.now() 
print('시작:', now)

# 'uid'별로 그룹화하여 'subject_traditional'의 고유한 개수를 계산
# 그리고 그 결과를 'cate_n'이라는 새로운 컬럼에 저장
df1['cate_n'] = df1.groupby('refs_uid')['subject_traditional'].transform(lambda x: x.nunique())
#df2['cate_n'] = df2.groupby('refs_uid')['subject_traditional'].transform(lambda x: x.nunique())
#df3['cate_n'] = df3.groupby('refs_uid')['subject_traditional'].transform(lambda x: x.nunique())

now = datetime.datetime.now()
print('종료:', now)

now = datetime.datetime.now() 
print('시작:', now)

# 'uid'별로 그룹화하여 'subject_traditional'의 고유한 개수를 계산
# 그리고 그 결과를 'cate_n'이라는 새로운 컬럼에 저장
#df1['cate_n'] = df1.groupby('refs_uid')['subject_traditional'].transform(lambda x: x.nunique())
#df2['cate_n'] = df2.groupby('refs_uid')['subject_traditional'].transform(lambda x: x.nunique())
df3['cate_n'] = df3.groupby('refs_uid')['subject_traditional'].transform(lambda x: x.nunique())

now = datetime.datetime.now()
print('종료:', now)

now = datetime.datetime.now() 
print('시작:', now)

# 'uid'별로 그룹화하여 'subject_traditional'의 고유한 개수를 계산
# 그리고 그 결과를 'cate_n'이라는 새로운 컬럼에 저장
#df1['cate_n'] = df1.groupby('refs_uid')['subject_traditional'].transform(lambda x: x.nunique())
df2['cate_n'] = df2.groupby('refs_uid')['subject_traditional'].transform(lambda x: x.nunique())
#df3['cate_n'] = df3.groupby('refs_uid')['subject_traditional'].transform(lambda x: x.nunique())

now = datetime.datetime.now()
print('종료:', now)

df1

unique_values = df1['cate_n'].value_counts()
print(unique_values)

# 유일성 체크: 'Article'과 'Review'의 빈도수 확인
unique_values22 = df2['cate_n'].value_counts()
print(unique_values22)

# 유일성 체크: 'Article'과 'Review'의 빈도수 확인
unique_values3 = df3['cate_n'].value_counts()
print(unique_values3)

unique_values = df1['cate_n'].unique()
print(unique_values)


df1.rename(columns={'refs_n' : 'ref_n'},inplace=True)

df3.rename(columns={'refs_n' : 'ref_n'},inplace=True)

df2.rename(columns={'refs_n' : 'ref_n'},inplace=True)

#컬럼 순서를 변경하고 싶은 순서대로 지정
new_order = ['uid' , 'cluster_id',  'pubyear', 'ref_n', 'refs_uid', 'cate_n', 'subject_traditional', 'within_core']

# 새 순서로 DataFrame을 재구성
df1_reordered = df1[new_order]
#df2_reordered = df2[new_order]
#df3_reordered = df3[new_order]

#컬럼 순서를 변경하고 싶은 순서대로 지정
new_order = ['uid' , 'cluster_id',  'pubyear', 'ref_n', 'refs_uid', 'cate_n', 'subject_traditional', 'within_core']

# 새 순서로 DataFrame을 재구성
#df1_reordered = df1[new_order]
#df2_reordered = df2[new_order]
df3_reordered = df3[new_order]

#컬럼 순서를 변경하고 싶은 순서대로 지정
new_order = ['uid' , 'cluster_id',  'pubyear', 'ref_n', 'refs_uid', 'cate_n', 'subject_traditional', 'within_core']

# 새 순서로 DataFrame을 재구성
#df1_reordered = df1[new_order]
#df2_reordered = df2[new_order]
df2_reordered = df2[new_order]

df1_reordered

df3_reordered

df2_reordered

within_core조사

now = datetime.datetime.now() 
print('시작:', now)
within_core_counts1 = df1_reordered['within_core'].value_counts()

# 결과 출력
print('df1')
print(within_core_counts1)

시작: 2024-03-04 16:47:32.936105
df1
within_core
0    339005389
Name: count, dtype: int64

now = datetime.datetime.now() 
print('시작:', now)
within_core_counts222 = df2_reordered['within_core'].value_counts()

# 결과 출력
print('df2')
print(within_core_counts222)

now = datetime.datetime.now() 
print('시작:', now)
within_core_counts33 = df3_reordered['within_core'].value_counts()

# 결과 출력
print('df3')
print(within_core_counts33)

저장

df1_reordered.to_csv('./final/four/four_1.txt', sep='\t', index=False)

df3_reordered.to_csv('./final/four/four_3.txt', sep='\t', index=False)

df2_reordered.to_csv('./final/four/four_2.txt', sep='\t', index=False)

df1_reordered

df2_reordered

## 0, 1, 2 불러와서 subject_traditional 전처리   
각 df 카테고리 개수 합집합하여 254개 나오는지 검증  
각각 저장도 하기  
최종결과물로 파일 병합하여 저장  


# reading 
now = datetime.datetime.now() 
print('시작:', now)
four_1 = pd.read_csv('./final/four/four_1.txt', sep='\t')

now = datetime.datetime.now()
print('종료:', now)

# reading 
now = datetime.datetime.now() 
print('시작:', now)
four_2 = pd.read_csv('./final/four/four_2.txt', sep='\t')

now = datetime.datetime.now()
print('종료:', now)

# reading 
now = datetime.datetime.now() 
print('시작:', now)
four_3 = pd.read_csv('./final/four/four_3.txt', sep='\t')

now = datetime.datetime.now()
print('종료:', now)

GREEN & SUSTAINABLE SCIENCE & TECHNOLOGY  
→  
Green & Sustainable Science & Technoology 

# subject_traditional이 'GREEN & SUSTAINABLE SCIENCE & TECHNOLOGY'인 행 추출
result_df_1_1 = four_1[four_1['subject_traditional'] == 'GREEN & SUSTAINABLE SCIENCE & TECHNOLOGY']
result_df_1_2 = four_1[four_1['subject_traditional'] == 'Green & Sustainable Science & Technology']
print(len(result_df_1_1))
print(len(result_df_1_2))

# subject_traditional이 'GREEN & SUSTAINABLE SCIENCE & TECHNOLOGY'인 행 추출
result_df_2_1 = four_2[four_2['subject_traditional'] == 'GREEN & SUSTAINABLE SCIENCE & TECHNOLOGY']
result_df_2_2 = four_2[four_2['subject_traditional'] == 'Green & Sustainable Science & Technology']
print(len(result_df_2_1))
print(len(result_df_2_2))

# subject_traditional이 'GREEN & SUSTAINABLE SCIENCE & TECHNOLOGY'인 행 추출
result_df_3_1 = four_3[four_3['subject_traditional'] == 'GREEN & SUSTAINABLE SCIENCE & TECHNOLOGY']
result_df_3_2 = four_3[four_3['subject_traditional'] == 'Green & Sustainable Science & Technology']
print(len(result_df_3_1))
print(len(result_df_3_2))

WOS Category 전처리

# 'subject_traditional' 컬럼 값 변경
four_1.loc[four_1['subject_traditional'] == 'GREEN & SUSTAINABLE SCIENCE & TECHNOLOGY', 'subject_traditional'] = 'Green & Sustainable Science & Technology'

# 'subject_traditional' 컬럼 값 변경
four_2.loc[four_2['subject_traditional'] == 'GREEN & SUSTAINABLE SCIENCE & TECHNOLOGY', 'subject_traditional'] = 'Green & Sustainable Science & Technology'

# 'subject_traditional' 컬럼 값 변경
four_3.loc[four_3['subject_traditional'] == 'GREEN & SUSTAINABLE SCIENCE & TECHNOLOGY', 'subject_traditional'] = 'Green & Sustainable Science & Technology'

# subject_traditional이 'GREEN & SUSTAINABLE SCIENCE & TECHNOLOGY'인 행 추출
result_df_1_1 = four_1[four_1['subject_traditional'] == 'GREEN & SUSTAINABLE SCIENCE & TECHNOLOGY']
result_df_1_2 = four_1[four_1['subject_traditional'] == 'Green & Sustainable Science & Technology']
print(len(result_df_1_1))
print(len(result_df_1_2))

# subject_traditional이 'GREEN & SUSTAINABLE SCIENCE & TECHNOLOGY'인 행 추출
result_df_2_1 = four_2[four_2['subject_traditional'] == 'GREEN & SUSTAINABLE SCIENCE & TECHNOLOGY']
result_df_2_2 = four_2[four_2['subject_traditional'] == 'Green & Sustainable Science & Technology']
print(len(result_df_2_1))
print(len(result_df_2_2))

# subject_traditional이 'GREEN & SUSTAINABLE SCIENCE & TECHNOLOGY'인 행 추출
result_df_3_1 = four_3[four_3['subject_traditional'] == 'GREEN & SUSTAINABLE SCIENCE & TECHNOLOGY']
result_df_3_2 = four_3[four_3['subject_traditional'] == 'Green & Sustainable Science & Technology']
print(len(result_df_3_1))
print(len(result_df_3_2))

Category 개수 조사

now = datetime.datetime.now() 
print('시작:', now)

# 'subject_traditional' 컬럼의 고유한 값들을 찾습니다.


unique_subjects1 = four_1['subject_traditional'].unique()
unique_subjects2 = four_2['subject_traditional'].unique()
unique_subjects3 = four_3['subject_traditional'].unique()


# 'subject_traditional' 컬럼의 고유한 값들로 이루어진 세트 생성
set_unique_subjects1 = set(unique_subjects1)
set_unique_subjects2 = set(unique_subjects2)
set_unique_subjects3 = set(unique_subjects3)
print('num_set_df1 : ')
print(len(set_unique_subjects1))
print('num_set_df1 : ')
print(len(set_unique_subjects2))
print('num_set_df1 : ')
print(len(set_unique_subjects3))


# 세 세트의 합집합 구하기
union_set = set_unique_subjects1 | set_unique_subjects2 | set_unique_subjects3


# 합집합의 요소 개수 세기
union_count = len(union_set)



# 254개인지 확인
if union_count == 254:
    print("카테고리 개수가 정확히 254개입니다.")
else:
    print(f"카테고리 개수가 예상과 다릅니다: {union_count}개")
    
    
now = datetime.datetime.now()
print('종료:', now)

print(set_unique_subjects1 - set_unique_subjects2)
print(set_unique_subjects1 - set_unique_subjects3)

import pickle
# Pickle 파일에서 리스트 불러오기
with open('./mid/unique_subjects.pkl', 'rb') as file:
    loaded_unique_subjects = pickle.load(file)

loaded_unique_subjects = set(loaded_unique_subjects)
print(set_unique_subjects1 - loaded_unique_subjects)

print(len(loaded_unique_subjects))

WOS Category에 속하지 않는{'Planning & Development'} 행은 드롭.

# subject_traditional이 'GREEN & SUSTAINABLE SCIENCE & TECHNOLOGY'인 행 추출
result1 = four_1[four_1['subject_traditional'] == 'Planning & Development']
result2 = four_2[four_2['subject_traditional'] == 'Planning & Development']
result3 = four_3[four_3['subject_traditional'] == 'Planning & Development']

print(len(result1))
print(len(result2))
print(len(result3))

# 'subject_traditional' 컬럼 값이 'Planning & Development'인 행 드롭
four_1 = four_1.drop(four_1[four_1['subject_traditional'] == 'Planning & Development'].index)

now = datetime.datetime.now() 
print('시작:', now)

# 'subject_traditional' 컬럼의 고유한 값들을 찾습니다.


unique_subjects1 = four_1['subject_traditional'].unique()
unique_subjects2 = four_2['subject_traditional'].unique()
unique_subjects3 = four_3['subject_traditional'].unique()


# 'subject_traditional' 컬럼의 고유한 값들로 이루어진 세트 생성
set_unique_subjects1 = set(unique_subjects1)
set_unique_subjects2 = set(unique_subjects2)
set_unique_subjects3 = set(unique_subjects3)
print('num_set_df1 : ')
print(len(set_unique_subjects1))
print('num_set_df1 : ')
print(len(set_unique_subjects2))
print('num_set_df1 : ')
print(len(set_unique_subjects3))


# 세 세트의 합집합 구하기
union_set = set_unique_subjects1 | set_unique_subjects2 | set_unique_subjects3


# 합집합의 요소 개수 세기
union_count = len(union_set)



# 254개인지 확인
if union_count == 254:
    print("카테고리 개수가 정확히 254개입니다.")
else:
    print(f"카테고리 개수가 예상과 다릅니다: {union_count}개")
    
    
now = datetime.datetime.now()
print('종료:', now)

print(loaded_unique_subjects- union_set)
print(union_set - loaded_unique_subjects)

중간 저장

four_1.to_csv('./final/four/final_four_1.txt', sep='\t', index=False)

four_2.to_csv('./final/four/final_four_2.txt', sep='\t', index=False)

four_3.to_csv('./final/four/final_four_3.txt', sep='\t', index=False)

불러와서 병합하여 저장

# reading 
now = datetime.datetime.now() 
print('시작:', now)
four_1 = pd.read_csv('./final/four/final_four_1.txt', sep='\t')
four_2 = pd.read_csv('./final/four/final_four_2.txt', sep='\t')
four_3 = pd.read_csv('./final/four/final_four_3.txt', sep='\t')

now = datetime.datetime.now()
print('종료:', now)

드롭한 행이 있으니 cate_n 다시

now = datetime.datetime.now() 
print('시작:', now)

four_1.drop(['cate_n'], axis=1, inplace=True)

four_1['cate_n'] = four_1.groupby('refs_uid')['subject_traditional'].transform(lambda x: x.nunique())

#컬럼 순서를 변경하고 싶은 순서대로 지정
new_order = ['uid' , 'cluster_id',  'pubyear', 'ref_n', 'refs_uid', 'cate_n', 'subject_traditional', 'within_core']

four_11 = four_1[new_order]
now = datetime.datetime.now()
print('종료:', now)

now = datetime.datetime.now() 
print('시작:', now)

four_2.drop(['cate_n'], axis=1, inplace=True)

four_2['cate_n'] = four_2.groupby('refs_uid')['subject_traditional'].transform(lambda x: x.nunique())

#컬럼 순서를 변경하고 싶은 순서대로 지정
new_order = ['uid' , 'cluster_id',  'pubyear', 'ref_n', 'refs_uid', 'cate_n', 'subject_traditional', 'within_core']

four_22 = four_2[new_order]
now = datetime.datetime.now()
print('종료:', now)

now = datetime.datetime.now() 
print('시작:', now)

four_3.drop(['cate_n'], axis=1, inplace=True)

four_3['cate_n'] = four_3.groupby('refs_uid')['subject_traditional'].transform(lambda x: x.nunique())

#컬럼 순서를 변경하고 싶은 순서대로 지정
new_order = ['uid' , 'cluster_id',  'pubyear', 'ref_n', 'refs_uid', 'cate_n', 'subject_traditional', 'within_core']

four_33 = four_3[new_order]
now = datetime.datetime.now()
print('종료:', now)

four_11.to_csv('./final/four/final_four_11.txt', sep='\t', index=False)

four_22.to_csv('./final/four/final_four_22.txt', sep='\t', index=False)

four_33.to_csv('./final/four/final_four_33.txt', sep='\t', index=False)


chunk_files = ['./final/four/final_four_11.txt','./final/four/final_four_22.txt', './final/four/final_four_33.txt' ]

# 병합된 파일 생성
output_file = './final/four.txt'

with open(output_file, 'w', newline='') as f_out:
    for i, file in enumerate(chunk_files):
        with open(file, 'r', newline='') as f_in:
            if i == 0:
                # 첫 번째 파일의 경우, 모든 내용을 복사
                f_out.write(f_in.read())
            else:
                # 두 번째 파일부터는 첫 줄(헤더)을 건너뛰고 나머지 내용을 복사
                next(f_in)  # 첫 줄(헤더) 건너뛰기
                f_out.write(f_in.read())

# =================================

최종파일 불러와서 검사는 메모리 에러..

# reading 
now = datetime.datetime.now() 
print('시작:', now)

four_ = pd.read_csv('./final/four.txt', sep='\t')


now = datetime.datetime.now()
print('종료:', now)

now = datetime.datetime.now() 
print('시작:', now)

# 'subject_traditional' 컬럼의 고유한 값들을 찾습니다.


unique_subjects = four['subject_traditional'].unique()


# 합집합의 요소 개수 세기
final_count = len(unique_subjects)



# 254개인지 확인
if final_count == 254:
    print("카테고리 개수가 정확히 254개입니다.")
else:
    print(f"카테고리 개수가 예상과 다릅니다: {final_count}개")
    
    
now = datetime.datetime.now()
print('종료:', now)
