==> 5_all_year.ipynb <==
# import modules
import pymysql
from sqlalchemy import create_engine
import pandas as pd
from pandas.io import sql
import datetime
import pandas as pd
import pandasql as ps

first는 leiden_cluster_biblio_info를 뜻합니다.(맨 처음 완성한 데이터셋)
여기서 필요한 컬럼만 뽑습니다.
uid 유니크를 검사합니다.

# reading 
now = datetime.datetime.now() 
print('시작:', now)
first = pd.read_csv('./final/first.txt', sep='\t')

now = datetime.datetime.now()
print('종료:', now)

selected_first = first[['uid', 'cluster_id', 'pubyear2']]


selected_first = selected_first.rename(columns={'pubyear2': 'pubyear'})


DB의 category_subject 테이블에서 subject_ascatype이 'traditional'인 것만 추출하여 저장한후, 그 파일을 불러옵니다.

# reading 
now = datetime.datetime.now() 
print('시작:', now)
sub =  pd.read_csv('./dict/category_subject.txt', sep='\t')
now = datetime.datetime.now()
print('종료:', now)

selected_first

sub

now = datetime.datetime.now() 
print('시작:', now)
# 청크와 다른 데이터셋 조인
five_result1 = pd.merge(selected_first, sub, on='uid', how='left')
#result_chunk = pd.merge(df_chunk, third, left_on='uid', right_on='refs_uid', how='inner')
now = datetime.datetime.now()
print('종료:', now)

#sorted_second_result3
# 각 컬럼별 null 값의 개수 계산
null_counts1 = five_result1.isnull().sum()

# 결과 출력
print(null_counts1)

# uid 컬럼의 고유한 값들의 수와 데이터 프레임의 행 수 비교
is_unique2 = len(selected_first['uid'].unique()) == len(selected_first)
print("UID 컬럼은 유일한가?", is_unique2)

결측치 체크를 한 후, 결측치를 드롭합니다.
cate_n과 subject_traditional 의 21821개의 null 값이 존재하여, 해당 행을 드롭했습니다.

df_cleaned = five_result1.dropna()


중간저장

now = datetime.datetime.now() 
print('시작:', now)
df_cleaned.to_csv('./mid/uid_sub.txt', sep='\t', index=False)
now = datetime.datetime.now()
print('종료:', now)

카테고리 유니크 개수를 조사합니다.
총 254개의 카테고리 개수가 나왔습니다.

unique_subjects = df_cleaned['subject_traditional'].unique()
print(len(unique_subjects))

254개의 카테고리 저장

나중에 wide 버전을 만들때 쓰일 카데고리 리스트를 pickle 파일로 미리 저장해놓습니다.

고유값들을 파일로 저장하고, 이를 나중에 다시 불러와서 DataFrame의 컬럼으로 사용하려는 목적이라면, 텍스트 파일 대신에 Python의 리스트 형태로 직접 저장하고 불러오는 방법이 더 효율적일 수 있습니다. 이 방법은 데이터를 구조화된 형태로 저장하고, 나중에 정확하게 같은 형태로 불러올 수 있게 해줍니다.

Pickle 파일로 저장하고 불러오기
Pickle은 Python 객체를 바이트 스트림으로 직렬화하고, 이를 파일로 저장하여 나중에 다시 불러올 수 있게 하는 라이브러리입니다. 이 방법을 사용하면, 고유값들을 리스트 형태로 직접 저장할 수 있으며, 나중에 이 리스트를 불러와서 즉시 사용할 수 있습니다.



import pickle

# 고유값 리스트 생성
unique_subjects_list = df_cleaned['subject_traditional'].unique().tolist()

# Pickle 파일로 저장
with open('./mid/unique_subjects.pkl', 'wb') as file:
    pickle.dump(unique_subjects_list, file)


#import pickle
# Pickle 파일에서 리스트 불러오기
#with open('./mid/unique_subjects.pkl', 'rb') as file:
#    loaded_unique_subjects = pickle.load(file)

# 이제 loaded_unique_subjects 리스트를 사용하여 DataFrame의 컬럼을 생성할 수 있음


# reading 
now = datetime.datetime.now() 
print('시작:', now)
uid_sub = pd.read_csv('./mid/uid_sub.txt', sep='\t')

now = datetime.datetime.now()
print('종료:', now)

uid_sub

일대일 대응 체크

# uid 컬럼의 고유한 값들의 수와 데이터 프레임의 행 수 비교
is_unique8 = len(uid_sub['uid'].unique()) == len(uid_sub)
print("UID 컬럼은 유일한가?", is_unique8)

uid별 subject_category가 몇개인지 나타내주는 sub_n 컬럼을 생성합니다.

uid_sub['sub_n'] = uid_sub.groupby('uid')['subject_traditional'].transform(lambda x: x.nunique())
#df2['cate_n'] = df2.groupby('refs_uid')['subject_traditional'].transform(lambda x: x.nunique())

uid_sub

now = datetime.datetime.now() 
print('시작:', now)

uid_sub.to_csv('./final/five/temp/mid_result_of_five.txt', sep='\t', index=False)

now = datetime.datetime.now()
print('종료:', now)

check point

five = uid_sub[['cluster_id', 'pubyear', 'uid' ,'subject_traditional', 'sub_n']]



sorted_five = five.sort_values(by=['cluster_id', 'pubyear'], ascending=[True, True])


c_id_counts = sorted_five['cluster_id'].value_counts()
print(c_id_counts)

dd_counts = sorted_five['sub_n'].value_counts()
print(dd_counts)

아래의 컬럼을 생성해줍니다.

### w = 1 / (sub_n)
### weight_cate : 한 클러스터 안에서 카테고리별 각각 w를 합산한 값
### norm_weight_cate : weight_cate를 unique한 uid 개수로 나눈 값
(weight_cate, norm_weight_cate 계산시 기준 : all_year는 클러스터별로, by_year는 클러스터별+년도별로 계산)

작업시 느낀점 기록 : weight_cate와 norm_weight_cate를 계산하고, 원래 데이터프레임에 merge해주는데, 이 작업은 불필요한 것 같습니다.

# Assuming sorted_five is your dataframe and it already exists with a column named 'sub_n'
# Example to add a new column 'w' where w = 1 / sub_n

sorted_five['w'] = 1 / sorted_five['sub_n']

sorted_five

# 클러스터당 uid unique 개수 카운트
# df_example 데이터프레임을 사용하여 cluster_id 별로 uid의 유니크한 개수를 카운트합니다.
uid_counts_per_cluster = sorted_five.groupby('cluster_id')['uid'].nunique().reset_index(name='unique_uid_count')

# 원래 df_example 데이터프레임에 uid_counts_per_cluster 데이터프레임을 merge하여,
# 각 row에 대한 cluster_id별 unique uid 개수를 추가합니다.
sorted_five = pd.merge(sorted_five, uid_counts_per_cluster, on='cluster_id')





# cluster_id와 subject_traditional로 groupby 하고, w 컬럼의 합계를 계산 후,
# 이를 원본 df_example에 merge하여 각 row에 대한 weight_cate 값을 추가합니다.
grouped_five = sorted_five.groupby(['cluster_id', 'subject_traditional'])['w'].sum().reset_index(name='weight_cate')

# 원본 데이터프레임에 weight_cate 값을 merge합니다.
five_result1 = pd.merge(sorted_five, grouped_five, on=['cluster_id', 'subject_traditional'])




five_result1

five_result1['norm_weight_cate'] = five_result1['weight_cate'] / sorted_five['unique_uid_count']

five_result1

now = datetime.datetime.now() 
print('시작:', now)

five_result1.to_csv('./final/five/temp/mid_result_of_five2.txt', sep='\t', index=False)

now = datetime.datetime.now()
print('종료:', now)

1. final_result1에서 cluster_id와 subject_tra를 groupby하기  
2.필요 컬럼만 추출하여 저장 = 5-1번


# 'five_result1' 데이터프레임에서 'cluster_id'와 'subject_traditional'로 groupby한 후, 각 그룹의 첫 번째 행을 선택
grouped_df = five_result1.groupby(['cluster_id', 'subject_traditional']).first().reset_index()

grouped_df

now = datetime.datetime.now() 
print('시작:', now)

five_result1.to_csv('./final/five/temp/mid_result_of_five3.txt', sep='\t', index=False)

now = datetime.datetime.now()
print('종료:', now)

# Selecting specific columns from the grouped DataFrame
selected_columns_df = grouped_df[['cluster_id', 'subject_traditional', 'weight_cate', 'norm_weight_cate']]



now = datetime.datetime.now() 
print('시작:', now)

selected_columns_df.to_csv('./final/five/temp/mid_result_of_five4.txt', sep='\t', index=False)

now = datetime.datetime.now()
print('종료:', now)

# reading 
now = datetime.datetime.now() 
print('시작:', now)
first = pd.read_csv('./final/five/temp/mid_result_of_five4.txt', sep='\t')

now = datetime.datetime.now()
print('종료:', now)

first

unique_counts = first.groupby('cluster_id')['subject_traditional'].nunique()

# 결과 출력
print(unique_counts)

import pickle
# Pickle 파일에서 리스트 불러오기
with open('./mid/unique_subjects.pkl', 'rb') as file:
    loaded_unique_subjects = pickle.load(file)

# 이제 loaded_unique_subjects 리스트를 사용하여 DataFrame의 컬럼을 생성할 수 있음


print(len(loaded_unique_subjects))

loaded_unique_subjects

1. 254 * 4215 뼈대 만들기  
2. left join하기 - 결측값은 0으로 채우기  

import pandas as pd
import numpy as np

# cluster_id 생성: 1부터 4215까지
cluster_ids = np.arange(1, 4216)

print(cluster_ids)

subject_traditional = loaded_unique_subjects

print(len(subject_traditional))

# DataFrame 생성
df = pd.DataFrame({
    'cluster_id': np.repeat(cluster_ids, len(subject_traditional)),
    'subject_traditional': np.tile(subject_traditional, len(cluster_ids))
})

df

unique_counts = df.groupby('cluster_id')['subject_traditional'].nunique()

# 결과 출력
print(unique_counts)


now = datetime.datetime.now() 
print('시작:', now)

result= pd.merge(df, first, on=['cluster_id','subject_traditional'], how='left')

now = datetime.datetime.now()
print('종료:', now)

result

result.fillna(0, inplace=True)

result

# 'cluster_id'를 우선적으로, 그리고 'subject_traditional'을 알파벳 순으로 정렬
result = result.sort_values(by=['cluster_id', 'subject_traditional'])


result.to_csv('./final/five/temp/mid_result_of_five5.txt', sep='\t', index=False)

result


# pivot_table을 사용하여 wide 형식의 데이터프레임 생성
pivot_df = result.pivot_table(index='cluster_id', columns='subject_traditional', values='norm_weight_cate', fill_value=0)

# 컬럼 순서 조정을 위한 준비: 'pubyear' 컬럼 추가
pivot_df['pubyear'] = 'all'

# 원하는 컬럼 순서 생성: 'cluster_id' 다음에 'pubyear'가 오도록
new_columns_order = ['pubyear'] + [col for col in pivot_df.columns if col not in [ 'pubyear']]

# DataFrame의 컬럼을 새 순서로 재배열
pivot_df_1 = pivot_df[new_columns_order]

# 인덱스에 이름을 지정
pivot_df_1.index.name = 'cluster_id'


pivot_df_1.to_csv('./final/five/temp/mid_result_of_five6.txt', sep='\t', index=True)
