==> 5_check.ipynb <==
## 5의 norm_wight_cate의 총계를 검사합니다. 

### 이슈 : norn_weight_cate의 총합을 구하니 1이 아닌 값이 나오는 이상치가 존재한다.  
하지만 weight_cate를 unique한 uid개수로 나누면서 소수점이 생기고,이 소수점이 무제한 기록이 안되고, 거기에서 총합을 구하니 1의 값이 아닌 이상치가 존재하게 되는거다. 하지만 1에 아주 근접한 값이고, 소수점상의 오류이므로 이번은 넘어간다. 총계를 1로 만들고 싶으면..........(생략..)

# import modules
import pymysql
from sqlalchemy import create_engine
import pandas as pd
from pandas.io import sql
import datetime
import pandas as pd
import pandasql as ps

# reading 
now = datetime.datetime.now() 
print('시작:', now)
five_all = pd.read_csv('./final/five/cluster_cate_weight_publevel/five_allyear.txt', sep='\t')
five_by_year = pd.read_csv('./final/five/cluster_cate_weight_publevel/five_by_year.txt', sep='\t')
now = datetime.datetime.now()
print('종료:', now)

check1= five_all.groupby(['cluster_id'])['norm_weight_cate'].sum().reset_index(name='sum_weight')

check1

check2 = five_by_year.groupby(['cluster_id', 'pubyear'])['norm_weight_cate'].sum().reset_index(name='sum_weight')

check2

check1.to_excel('./check/five_all.xlsx', index=False)
check2.to_excel('./check/five_by_year.xlsx', index=False)

# 컬럼에 포함된 고유한 값들과 그 빈도수를 확인
unique_values_counts_sum_weight1 = check1['sum_weight'].value_counts()
print(unique_values_counts_sum_weight1)

#컬럼에 포함된 고유한 값들과 그 빈도수를 확인
unique_values_counts_sum_weight2 = check2['sum_weight'].value_counts()
print(unique_values_counts_sum_weight2)

five_all
five_all_wide
five_by_year
five_byyear_wide
